Make database interface.
