+++ ui/ui.go
@@ -59,6 +59,8 @@ func Start(cfg Config, listener net.Listener) {
 }
 
 func healthcheckHandler(cfg Config) http.Handler {
+	db := db.Pqdb{}
+
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		_, err := db.GetAllTasks(cfg.DB)
